backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.35
psycopg[binary]==3.1.13
alembic==1.13.0

# Redis and Caching
redis==5.0.1
hiredis==2.2.3

# Background Tasks
celery==5.3.4

# AI/ML Libraries
langchain==0.0.340
openai==1.3.5
pinecone-client==2.2.4

# Document Processing
pypdf==3.17.1
python-docx==1.1.0
beautifulsoup4==4.12.2
requests==2.31.0

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Monitoring and Logging
sentry-sdk[fastapi]==1.38.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Production
gunicorn==21.2.0

# Optional packages (install separately if needed)
# ollama==0.1.8  # For local LLM
# unstructured==0.11.6  # For advanced document processing
# structlog==23.2.0  # For structured logging